REF: moved ErrorProvider to Dialog
